AffiliateAndPurchase/requirements.txt
pandas>=2.0
numpy>=1.24
pymongo>=4.0
sshtunnel>=0.4
python-dotenv>=1.0
requests>=2.28
tabulate>=0.9
